chore: minor refactor with ProtocolGenerators and encode/decode
