+++ codegen/smithy-aws-swift-codegen/src/main/kotlin/software/amazon/smithy/aws/swift/codegen/AWSHttpRequestJsonEncoder.kt
@@ -1,0 +1,12 @@
+package software.amazon.smithy.aws.swift.codegen
+
+import software.amazon.smithy.swift.codegen.SwiftWriter
+import software.amazon.smithy.swift.codegen.integration.HttpRequestEncoder
+
+class AWSHttpRequestJsonEncoder(
+    requestEncoderOptions: MutableMap<String, String> = mutableMapOf()
+) : HttpRequestEncoder("JSONEncoder", requestEncoderOptions) {
+    override fun renderInitialization(writer: SwiftWriter, nameOfConfigObject: String) {
+        writer.write("self.encoder = \$L.encoder ?? \$L", nameOfConfigObject, name)
+    }
+}
