+++ codegen/smithy-aws-swift-codegen/src/main/kotlin/software/amazon/smithy/aws/swift/codegen/AWSHttpResponseJsonDecoder.kt
@@ -1,0 +1,12 @@
+package software.amazon.smithy.aws.swift.codegen
+
+import software.amazon.smithy.swift.codegen.SwiftWriter
+import software.amazon.smithy.swift.codegen.integration.HttpResponseDecoder
+
+class AWSHttpResponseJsonDecoder(
+    responseDecoderOptions: MutableMap<String, String> = mutableMapOf()
+) : HttpResponseDecoder("JSONDecoder", responseDecoderOptions) {
+    override fun renderInitialization(writer: SwiftWriter, nameOfConfigObject: String) {
+        writer.write("self.decoder = \$L.decoder ?? \$L", nameOfConfigObject, name)
+    }
+}
