+++ codegen/smithy-aws-swift-codegen/src/main/kotlin/software/amazon/smithy/aws/swift/codegen/awsjson/AwsJson1_0_ProtocolGenerator.kt
@@ -1,13 +1,13 @@
 package software.amazon.smithy.aws.swift.codegen.awsjson
 
-import software.amazon.smithy.aws.swift.codegen.JsonProtocolBase
+import software.amazon.smithy.aws.swift.codegen.AWSHttpBindingJsonProtocolGenerator
 import software.amazon.smithy.aws.traits.protocols.AwsJson1_0Trait
 import software.amazon.smithy.model.shapes.ShapeId
 import software.amazon.smithy.model.traits.TimestampFormatTrait
 import software.amazon.smithy.swift.codegen.integration.HttpBindingResolver
 import software.amazon.smithy.swift.codegen.integration.ProtocolGenerator
 
-open class AwsJson1_0_ProtocolGenerator : JsonProtocolBase() {
+open class AwsJson1_0_ProtocolGenerator : AWSHttpBindingJsonProtocolGenerator() {
     override val defaultContentType: String = "application/x-amz-json-1.0"
     override val defaultTimestampFormat: TimestampFormatTrait.Format = TimestampFormatTrait.Format.EPOCH_SECONDS
     override val protocol: ShapeId = AwsJson1_0Trait.ID
