+++ codegen/smithy-aws-swift-codegen/src/main/kotlin/software/amazon/smithy/aws/swift/codegen/restjson/AWSRestJson1ProtocolGenerator.kt
@@ -4,7 +4,7 @@
  */
 package software.amazon.smithy.aws.swift.codegen.restjson
 
-import software.amazon.smithy.aws.swift.codegen.JsonProtocolBase
+import software.amazon.smithy.aws.swift.codegen.AWSHttpBindingJsonProtocolGenerator
 import software.amazon.smithy.aws.traits.protocols.RestJson1Trait
 import software.amazon.smithy.model.shapes.ShapeId
 import software.amazon.smithy.model.traits.TimestampFormatTrait
@@ -13,10 +13,10 @@ import software.amazon.smithy.model.traits.TimestampFormatTrait
  * Handles generating the aws.rest-json protocol for services.
  *
  * @inheritDoc
- * @see JsonProtocolBase
+ * @see AWSHttpBindingJsonProtocolGenerator
  */
 
-class AWSRestJson1ProtocolGenerator : JsonProtocolBase() {
+class AWSRestJson1ProtocolGenerator : AWSHttpBindingJsonProtocolGenerator() {
     override val defaultContentType: String = "application/json"
     override val defaultTimestampFormat: TimestampFormatTrait.Format = TimestampFormatTrait.Format.EPOCH_SECONDS
     override val protocol: ShapeId = RestJson1Trait.ID
